+++ sd3_train.py
@@ -29,6 +29,7 @@ from library.sdxl_train_util import match_mixed_precision
 
 import library.train_util as train_util
 import library.logging_util as logging_util
+import library.loss as loss_util
 
 from library.utils import setup_logging, add_logging_arguments
 
@@ -849,8 +850,8 @@ def train(args):
                 #     1,
                 # )
                 # calculate loss
-                huber_c = train_util.get_huber_threshold_if_needed(args, timesteps, dummy_scheduler)
-                loss = train_util.conditional_loss(model_pred.float(), target.float(), args.loss_type, "none", huber_c)
+                huber_c = loss_util.get_huber_threshold_if_needed(args, timesteps, dummy_scheduler)
+                loss = loss_util.conditional_loss(model_pred.float(), target.float(), args.loss_type, "none", huber_c)
                 if args.masked_loss or ("alpha_masks" in batch and batch["alpha_masks"] is not None):
                     loss = apply_masked_loss(loss, batch)
                 loss = loss.mean([1, 2, 3])
